Fix: Address All Linting Errors From Your Feedback
I've applied fixes for all the linting errors and warnings you reported.

Here are the changes I made:
- `DataTable.jsx`: Removed an unused import of `saveDocument`.
- `ToastContext.jsx`:
  - Added `useEffect` to the React import to fix a 'not defined' error.
  - Added the `removeToast` function to the `useCallback` dependency array to fix a missing dependency warning.
- Fast Refresh Fixes: To resolve the 'Fast refresh only works when a file only exports components' warning and improve code structure:
  - The `Toast` component was extracted from `ToastContext.jsx` into its own file (`src/components/Toast.jsx`).
  - `ToastProvider` in `ToastContext.jsx` was changed to a default export.
  - `AuthProvider` in `AuthContext.jsx` was changed to a default export.
  - The import statement for `AuthProvider` in `main.jsx` was updated to match the default export.

These changes result in a cleaner, more correct codebase that adheres to React best practices.
