+++ src/contexts/ToastContext.jsx
@@ -1,4 +1,5 @@
-import React, { createContext, useContext, useState, useCallback } from 'react';
+import React, { createContext, useContext, useState, useCallback, useEffect } from 'react';
+import Toast from '../components/Toast'; // Import the new component
 
 const ToastContext = createContext();
 
@@ -6,8 +7,12 @@ export const useToast = () => useContext(ToastContext);
 
 let idCounter = 0;
 
-export const ToastProvider = ({ children }) => {
+const ToastProvider = ({ children }) => {
   const [toasts, setToasts] = useState([]);
+
+  const removeToast = useCallback((id) => {
+    setToasts(currentToasts => currentToasts.filter(toast => toast.id !== id));
+  }, []);
 
   const addToast = useCallback((message, type = 'info', duration = 4000) => {
     const id = idCounter++;
@@ -16,11 +21,7 @@ export const ToastProvider = ({ children }) => {
     setTimeout(() => {
       removeToast(id);
     }, duration);
-  }, []);
-
-  const removeToast = useCallback((id) => {
-    setToasts(currentToasts => currentToasts.filter(toast => toast.id !== id));
-  }, []);
+  }, [removeToast]); // Added missing dependency
 
   return (
     <ToastContext.Provider value={{ addToast }}>
@@ -34,32 +35,4 @@ export const ToastProvider = ({ children }) => {
   );
 };
 
-// The actual Toast component that renders the UI for a single toast.
-// Placing it here for simplicity as it's tightly coupled to the provider.
-const Toast = ({ message, type, onClose }) => {
-  const [show, setShow] = useState(false);
-
-  useEffect(() => {
-    // Animate in
-    setShow(true);
-
-    // The parent provider will handle the removal timeout.
-    // This effect is just for the entrance animation.
-  }, []);
-
-  const typeClasses = {
-    info: 'bg-blue-600',
-    success: 'bg-green-600',
-    error: 'bg-red-600',
-  };
-
-  return (
-    <div
-      className={`flex items-center px-5 py-3 rounded-lg text-white font-semibold shadow-lg transition-all duration-300 ease-in-out transform ${typeClasses[type] || 'bg-slate-700'} ${show ? 'translate-x-0 opacity-100' : 'translate-x-full opacity-0'}`}
-      role="alert"
-    >
-      <span>{message}</span>
-      <button onClick={onClose} className="ml-4 p-1 rounded-full hover:bg-white/20">&times;</button>
-    </div>
-  );
-};
+export default ToastProvider; // Use default export to solve Fast Refresh issue
