+++ src/main.jsx
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App.jsx';
 import './index.css';
-import { AuthProvider } from './contexts/AuthContext.jsx';
-import { ToastProvider } from './contexts/ToastContext.jsx';
+import AuthProvider from './contexts/AuthContext.jsx';
+import ToastProvider from './contexts/ToastContext.jsx';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
